Final Year Project/Unity Game Engine
Final year, Computing System Project, based around the fnctionality of the AI in Unity Game Engine.
The aim of the paper is to attempt to enhance the ability of the AI to use the enviroment, game scene, as a cover on its way to a target, without beeing spotted. Along with this task, the game performance is improved for a better game experience.